requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
groq
streamlit
python-dotenv
scikit-learn
